+++ app/service/mongodb/dao/user_dao.py
@@ -1,31 +1,26 @@
-from typing import Dict
-from app.service.mongodb.dao.base_dao import BaseDAO
-from app.service.mongodb.mongo_service import MongoService
+from datetime import datetime
+from pymongo.collection import Collection
+from app.telegram.model.user import User
 
-class UserDAO(BaseDAO):
-    """DAO for interacting with the user collection."""
+class UserDAO:
+    def __init__(self, collection: Collection):
+        self._collection = collection
 
-    def __init__(self, mongo_service: MongoService, db_alias: str, collection_name: str):
-        super().__init__(mongo_service, db_alias)
-        self.collection = self.mongo_service.get_collection(db_alias, collection_name)
+    async def get_user_by_id(self, user_id: int) -> User:
+        user_data = await self._collection.find_one({"user_id": user_id})
+        if user_data:
+            return User(**user_data)
+        return None
 
-    async def get_user_data(self, user_id: int) -> dict:
-        """Retrieves user data from MongoDB."""
-        result = await self.collection.find_one({"_id": user_id})
-        return result.get("data", {}) if result else {}
-
-    async def update_user_data(self, user_id: int, data: dict) -> None:
-        """Updates user data in MongoDB."""
-        await self.collection.update_one(
-            {"_id": user_id},
-            {"$set": {"data": data}},
+    async def add_user(self, user: User) -> None:
+        self._collection.update_one(
+            {"user_id": user.user_id},
+            {"$set": user.model_dump()},
             upsert=True
         )
 
-    async def get_all_user_data(self) -> Dict[int, Dict]:
-        """Retrieves all user data from MongoDB."""
-        user_data = {}
-        async for user in self.collection.find():
-            user_id = user["_id"]
-            user_data[user_id] = user["data"]
-        return user_data
+    async def update_user_activity(self, user_id: int) -> None:
+        self._collection.update_one(
+            {"user_id": user_id},
+            {"$set": {"last_active": datetime.utcnow().isoformat()}}
+        )