+++ app/telegram/bot_dispatcher.py
@@ -4,6 +4,7 @@ from app.telegram.middleware.register import MiddlewareRegister
 from app.telegram.handler.get_search_params.register import SearchParamsRegister
 from app.telegram.handler.start.register import StartRegister
 from app.telegram.handler.menu.register import MenuRegister
+from app.telegram.handler.subscription.register import SubscriptionRegister
 
 class BotDispatcher:
 
@@ -27,3 +28,4 @@ class BotDispatcher:
         StartRegister(self.dispatcher, self.router_factory)
         SearchParamsRegister(self.dispatcher, self.router_factory)
         MenuRegister(self.dispatcher, self.router_factory)
+        SubscriptionRegister(self.dispatcher, self.router_factory)